ui_automation/selenium/requirements.txt
selenium
webdriver-manager
python-dotenv
schedule